requirements.txt
streamlit
tensorflow
opencv-python
numpy
pillow
ultralytics
mtcnn
